restructure diretory
