+++ lib/ruby_sdk/client_with_context.rb
