+++ lib/ruby_sdk/context.rb
