+++ lib/ruby_sdk/event_source_client.rb
