+++ lib/ruby_sdk/feature_state.rb
