+++ lib/ruby_sdk/mod/handle_undefined_feature.rb
