+++ lib/ruby_sdk/strategy.rb
